Optimize syntax. Non capturing groups and short `solFuncCall`.
End context in `solFuncCall` seems to capture more than necessary.  As
it is a regex, it still can not tell, for a closing curvy brace or
parenthesis, if it is matching the opening one or not.  So it seems
wasteful to search for anything beyond the opening curvy brace or
parenthesis.

All the capturing groups were used only for grouping.  None of the
captured text was every used.  While it is a small optimization, I would
imagine, it could be noticeable when there are a lot of matches.

Similarly, for a sequence of words, it makes sense to match for the
beginning of word and the end of word only once, instead of repeatedly
for every word in the list that we are checking.

In a few places '\s*' was used where zero whitespace is not actually
allowed by the Solidity grammar.