sol{If,Loop}Block: solFuncBody groups in "contains="
Both "solIfBlock" and "solLoopBlock" may contain all the same groups as
the "solFuncBody".  As they all use "{" and "}" as the start and end
patterns respectively, they need to make sure other syntactic constructs
that also use curvy braces are recognised inside of both "if" and "loop"
statements.  Otherwise "if" and "loop" statements end prematurely.